chatbot_api/pyproject.toml
[project]
name = "chatbot_api"
version = "0.1"
dependencies = [
    "asyncio==3.4.3",
    "fastapi==0.109.0",
    "langchain==0.2.10",
    "langchain-openai==0.1.17",
    "langchain-community==0.2.9",
    "langchainhub==0.1.14",
    "neo4j==5.14.1",
    "numpy==1.26.2",
    "openai>=1.56.1",
    "opentelemetry-api==1.22.0",
    "pydantic==2.5.1",
    "uvicorn==0.25.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0,<3.0.0"
]

[project.optional-dependencies]
dev = ["black", "flake8"]